+++ examples/test/dynarray/test_dynarray_typedconst_badreassign.pas
@@ -1,0 +1,7 @@
+program TestDArrTCBadReassign;
+{ A BARE dynamic-array typed constant can't be reassigned. }
+const
+    X: array of integer = [1, 2];
+begin
+    X := [3, 4];
+end.