+++ examples/test/dynarray/test_dynarray_typedconst_badvalue.pas
@@ -1,0 +1,7 @@
+program TestDArrTCBadValue;
+{ A BARE dynamic-array typed constant's value must be an array literal
+  ('[...]') - nothing else, including 'nil', is accepted. }
+const
+    X: array of integer = nil;
+begin
+end.